+++ examples/frame_metalayers.c
@@ -16,6 +16,7 @@
  */
 
 #include <stdio.h>
+#include <string.h>
 #include <assert.h>
 #include <blosc.h>
 
@@ -84,12 +85,12 @@ int main() {
     ttotal = blosc_elapsed_secs(last, current);
     printf("Time for schunk -> frame: %.3g s, %.1f GB/s\n",
            ttotal, nbytes / (ttotal * GB));
-    printf("Frame length in memory: %lld bytes\n", frame_len);
+    printf("Frame length in memory: %ld bytes\n", (long)frame_len);
 
     // frame1 (in-memory) -> fileframe (on-disk)
     blosc_set_timestamp(&last);
     frame_len = blosc2_frame_to_file(&frame1, "frame_metalayers.b2frame");
-    printf("Frame length on disk: %lld bytes\n", frame_len);
+    printf("Frame length on disk: %ld bytes\n", (long)frame_len);
     blosc_set_timestamp(&current);
     ttotal = blosc_elapsed_secs(last, current);
     printf("Time for frame -> fileframe (simple_frame.b2frame): %.3g s, %.1f GB/s\n",
