+++ examples/frame_simple.c
@@ -93,12 +93,12 @@ int main() {
   ttotal = blosc_elapsed_secs(last, current);
   printf("Time for schunk -> frame: %.3g s, %.1f MB/s\n",
          ttotal, nbytes / (ttotal * MB));
-  printf("Frame length in memory: %lld bytes\n", frame_len);
+  printf("Frame length in memory: %ld bytes\n", (long)frame_len);
 
   // frame1 (in-memory) -> fileframe (on-disk)
   blosc_set_timestamp(&last);
   frame_len = blosc2_frame_to_file(&frame1, "frame_simple.b2frame");
-  printf("Frame length on disk: %lld bytes\n", frame_len);
+  printf("Frame length on disk: %ld bytes\n", (long)frame_len);
   blosc_set_timestamp(&current);
   ttotal = blosc_elapsed_secs(last, current);
   printf("Time for frame -> fileframe (frame_simple.b2frame): %.3g s, %.1f GB/s\n",
